rename: RootStackNavigator.tsx
